refactor(copyright): split token utils modules
