+++ src/copyright/detector/token_utils/filters.rs
@@ -1,80 +1,17 @@
 // SPDX-FileCopyrightText: Provenant contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use std::collections::{HashMap, HashSet};
+use std::collections::HashSet;
 use std::sync::LazyLock;
 
 use regex::Regex;
 
-use crate::copyright::refiner::{
-    is_junk_copyright, is_junk_holder, is_path_like_code_fragment, refine_author, refine_copyright,
-    refine_holder, refine_holder_in_copyright_context,
-};
+use super::normalize_whitespace;
+use crate::copyright::refiner::is_path_like_code_fragment;
 use crate::copyright::types::{
-    AuthorDetection, CopyrightDetection, HolderDetection, ParseNode, PosTag, Token, TreeLabel,
+    CopyrightDetection, HolderDetection, ParseNode, PosTag, Token, TreeLabel,
 };
 use crate::models::LineNumber;
-
-#[cfg(test)]
-#[path = "token_utils_test.rs"]
-mod tests;
-
-pub fn is_copyright_span_token(token: &Token) -> bool {
-    !matches!(token.tag, PosTag::EmptyLine | PosTag::Junk)
-}
-
-pub fn extract_original_author_additional_contributors(
-    tree: &[ParseNode],
-) -> Option<AuthorDetection> {
-    let all_leaves: Vec<&Token> = tree.iter().flat_map(collect_all_leaves).collect();
-    if all_leaves.is_empty() {
-        return None;
-    }
-
-    let mut has_original = false;
-    let mut has_author = false;
-    for t in &all_leaves {
-        let v = t
-            .value
-            .trim_matches(|c: char| c.is_ascii_punctuation())
-            .to_ascii_lowercase();
-        if v == "original" {
-            has_original = true;
-        } else if v == "author" {
-            has_author = true;
-        }
-    }
-    if !has_original || !has_author {
-        return None;
-    }
-
-    for (i, t) in all_leaves.iter().enumerate() {
-        let v = t
-            .value
-            .trim_matches(|c: char| c.is_ascii_punctuation())
-            .to_ascii_lowercase();
-        if v != "additional" {
-            continue;
-        }
-        let line = t.start_line;
-        for u in all_leaves.iter().skip(i + 1).take(6) {
-            if u.start_line != line {
-                break;
-            }
-            let uv = u
-                .value
-                .trim_matches(|c: char| c.is_ascii_punctuation())
-                .to_ascii_lowercase();
-            let is_contributors = u.tag == PosTag::Contributors || uv.starts_with("contributor");
-            if is_contributors {
-                let tokens: Vec<&Token> = vec![*t, *u];
-                return build_author_from_tokens(&tokens);
-            }
-        }
-    }
-
-    None
-}
 
 pub fn should_merge_following_copyright_clause(
     all_leaves: &[&Token],
@@ -170,30 +107,6 @@ pub fn should_merge_following_c_sign_after_year(
     })
 }
 
-pub fn is_author_span_token(token: &Token) -> bool {
-    !matches!(
-        token.tag,
-        PosTag::EmptyLine | PosTag::Junk | PosTag::Copy | PosTag::SpdxContrib
-    )
-}
-
-pub fn collect_all_leaves(node: &ParseNode) -> Vec<&Token> {
-    let mut result = Vec::new();
-    collect_all_leaves_inner(node, &mut result);
-    result
-}
-
-fn collect_all_leaves_inner<'a>(node: &'a ParseNode, result: &mut Vec<&'a Token>) {
-    match node {
-        ParseNode::Leaf(token) => result.push(token),
-        ParseNode::Tree { children, .. } => {
-            for child in children {
-                collect_all_leaves_inner(child, result);
-            }
-        }
-    }
-}
-
 pub fn apply_written_by_for_markers(
     group: &[(usize, String)],
     copyrights: &mut [CopyrightDetection],
@@ -272,215 +185,6 @@ pub fn restore_bare_holder_angle_emails(
     }
 }
 
-// ─── Detection builders from tree nodes ──────────────────────────────────────
-
-pub fn build_holder_from_node(
-    node: &ParseNode,
-    ignored_labels: &[TreeLabel],
-    ignored_pos_tags: &[PosTag],
-) -> Option<HolderDetection> {
-    let leaves = collect_holder_filtered_leaves(node, ignored_labels, ignored_pos_tags);
-    let filtered = strip_all_rights_reserved(leaves);
-    let allow_single_word_contributors = collect_all_leaves(node)
-        .iter()
-        .any(|t| matches!(t.tag, PosTag::Yr | PosTag::YrPlus | PosTag::BareYr));
-    build_holder_from_tokens(&filtered, allow_single_word_contributors)
-}
-
-pub fn build_holder_from_copyright_node(
-    node: &ParseNode,
-    ignored_labels: &[TreeLabel],
-    ignored_pos_tags: &[PosTag],
-) -> Option<HolderDetection> {
-    let all_leaves = collect_all_leaves(node);
-    let held_by_clause = all_leaves.len() >= 4
-        && all_leaves[0].tag == PosTag::Copy
-        && all_leaves[1].tag == PosTag::Is
-        && all_leaves[2].tag == PosTag::Held
-        && all_leaves[3].tag == PosTag::By;
-    if held_by_clause {
-        return None;
-    }
-
-    let copy_line = all_leaves
-        .iter()
-        .filter(|t| t.tag == PosTag::Copy && t.value.eq_ignore_ascii_case("copyright"))
-        .map(|t| t.start_line)
-        .min();
-
-    let keep_prefix_lines = copy_line
-        .map(|cl| signal_lines_before_copy_line(node, cl))
-        .unwrap_or_default();
-
-    let leaves = collect_holder_filtered_leaves(node, ignored_labels, ignored_pos_tags);
-    let mut filtered = strip_all_rights_reserved(leaves);
-    if let Some(copy_line) = copy_line {
-        filtered.retain(|t| {
-            t.start_line >= copy_line || keep_prefix_lines.contains(&t.start_line.get())
-        });
-    }
-
-    let allow_single_word_contributors = all_leaves
-        .iter()
-        .any(|t| matches!(t.tag, PosTag::Yr | PosTag::YrPlus | PosTag::BareYr));
-
-    build_holder_from_tokens(&filtered, allow_single_word_contributors)
-}
-
-pub fn signal_lines_before_copy_line(node: &ParseNode, copy_line: LineNumber) -> HashSet<usize> {
-    use std::collections::HashMap;
-
-    let mut by_line: HashMap<usize, Vec<&Token>> = HashMap::new();
-    for t in collect_all_leaves(node) {
-        if t.start_line < copy_line {
-            by_line.entry(t.start_line.get()).or_default().push(t);
-        }
-    }
-
-    let mut keep = HashSet::new();
-    for (line, tokens) in by_line {
-        let has_strong_signal = tokens.iter().any(|t| {
-            matches!(
-                t.tag,
-                PosTag::Yr
-                    | PosTag::YrPlus
-                    | PosTag::BareYr
-                    | PosTag::Copy
-                    | PosTag::Auth
-                    | PosTag::Auth2
-                    | PosTag::Auths
-                    | PosTag::AuthDot
-                    | PosTag::Maint
-                    | PosTag::Contributors
-                    | PosTag::Commit
-                    | PosTag::SpdxContrib
-            ) || t.value.eq_ignore_ascii_case("author")
-                || t.value.eq_ignore_ascii_case("authors")
-        });
-        if has_strong_signal {
-            keep.insert(line);
-            continue;
-        }
-
-        let clean: Vec<&Token> = tokens
-            .iter()
-            .copied()
-            .filter(|t| !matches!(t.tag, PosTag::Junk | PosTag::EmptyLine | PosTag::Parens))
-            .collect();
-        if clean.is_empty() {
-            continue;
-        }
-        if clean.len() > 3 {
-            continue;
-        }
-
-        let is_fragment = clean.iter().all(|t| {
-            let v = t.value.trim_matches(|c: char| !c.is_alphanumeric());
-            if v.is_empty() {
-                return false;
-            }
-            let lower = v.to_ascii_lowercase();
-            if matches!(
-                lower.as_str(),
-                "the" | "and" | "or" | "of" | "by" | "in" | "to"
-            ) {
-                return false;
-            }
-            v.chars().next().is_some_and(|c| c.is_ascii_uppercase())
-        });
-
-        if is_fragment {
-            keep.insert(line);
-        }
-    }
-
-    keep
-}
-
-pub fn build_author_from_node(node: &ParseNode) -> Option<AuthorDetection> {
-    let leaves = collect_filtered_leaves(
-        node,
-        &[TreeLabel::YrRange, TreeLabel::YrAnd],
-        super::NON_AUTHOR_POS_TAGS,
-    );
-    build_author_from_tokens(&leaves)
-}
-
-// ─── Detection builders from token slices ────────────────────────────────────
-
-pub fn build_copyright_from_tokens(tokens: &[&Token]) -> Option<CopyrightDetection> {
-    if tokens.is_empty() {
-        return None;
-    }
-    let node_string = normalized_tokens_to_string(tokens);
-    let refined = refine_copyright(&node_string)?;
-    if is_junk_copyright(&refined) {
-        return None;
-    }
-    Some(CopyrightDetection {
-        copyright: refined,
-        start_line: tokens
-            .first()
-            .map(|t| t.start_line)
-            .unwrap_or(LineNumber::ONE),
-        end_line: tokens
-            .last()
-            .map(|t| t.start_line)
-            .unwrap_or(LineNumber::ONE),
-    })
-}
-
-pub fn build_holder_from_tokens(
-    tokens: &[&Token],
-    allow_single_word_contributors: bool,
-) -> Option<HolderDetection> {
-    if tokens.is_empty() {
-        return None;
-    }
-    let node_string = normalized_tokens_to_string(tokens);
-    let refined = if allow_single_word_contributors {
-        refine_holder_in_copyright_context(&node_string)?
-    } else {
-        refine_holder(&node_string)?
-    };
-    if is_junk_copyright(&refined) || is_junk_holder(&refined) {
-        return None;
-    }
-    Some(HolderDetection {
-        holder: refined,
-        start_line: tokens
-            .first()
-            .map(|t| t.start_line)
-            .unwrap_or(LineNumber::ONE),
-        end_line: tokens
-            .last()
-            .map(|t| t.start_line)
-            .unwrap_or(LineNumber::ONE),
-    })
-}
-
-pub fn build_author_from_tokens(tokens: &[&Token]) -> Option<AuthorDetection> {
-    if tokens.is_empty() {
-        return None;
-    }
-    let node_string = normalized_tokens_to_string(tokens);
-    let refined = refine_author(&node_string)?;
-    if is_junk_copyright(&refined) {
-        return None;
-    }
-    Some(AuthorDetection {
-        author: refined,
-        start_line: tokens
-            .first()
-            .map(|t| t.start_line)
-            .unwrap_or(LineNumber::ONE),
-        end_line: tokens
-            .last()
-            .map(|t| t.start_line)
-            .unwrap_or(LineNumber::ONE),
-    })
-}
-
 pub fn looks_like_bad_generic_author_candidate(s: &str) -> bool {
     let trimmed = s.trim();
     if trimmed.is_empty() {
@@ -545,8 +249,6 @@ pub fn looks_like_bad_generic_author_candidate(s: &str) -> bool {
 
     false
 }
-
-// ─── Shared helpers ──────────────────────────────────────────────────────────
 
 pub fn collect_filtered_leaves<'a>(
     node: &'a ParseNode,
@@ -666,7 +368,7 @@ pub fn drop_path_fragment_holders_from_bare_c_code_lines(
 
 /// Tags whose filtering should cause adjacent commas to be considered orphaned.
 /// Only year-related tags: commas between years (e.g. "2006, 2007") become
-/// orphaned when the years are removed.  Email/URL commas are intentionally
+/// orphaned when the years are removed. Email/URL commas are intentionally
 /// excluded because they typically separate legitimate holder names
 /// (e.g. "Name <email>, Name2").
 pub const YEAR_LIKE_POS_TAGS: &[PosTag] = &[PosTag::Yr, PosTag::YrPlus, PosTag::BareYr];
@@ -966,45 +668,4 @@ pub fn is_copyright_of_header(span: &[&Token]) -> bool {
         .iter()
         .any(|t| t.tag == PosTag::Copy && t.value.eq_ignore_ascii_case("(c)"));
     !has_year && !has_c
-}
-
-pub fn normalized_tokens_to_string(tokens: &[&Token]) -> String {
-    let mut out = String::new();
-    let mut first = true;
-
-    for token in tokens {
-        for piece in token.value.split_whitespace() {
-            if !first {
-                out.push(' ');
-            }
-            out.push_str(piece);
-            first = false;
-        }
-    }
-
-    out
-}
-
-pub fn normalize_whitespace(s: &str) -> String {
-    s.split_whitespace().collect::<Vec<_>>().join(" ")
-}
-
-pub fn group_by<T, K>(items: Vec<T>, key_fn: impl Fn(&T) -> K) -> Vec<(K, Vec<T>)>
-where
-    K: std::hash::Hash + Eq + Clone,
-{
-    let mut order: Vec<K> = Vec::new();
-    let mut seen: HashSet<K> = HashSet::new();
-    let mut map: HashMap<K, Vec<T>> = HashMap::new();
-    for item in items {
-        let key = key_fn(&item);
-        if seen.insert(key.clone()) {
-            order.push(key.clone());
-        }
-        map.entry(key).or_default().push(item);
-    }
-    order
-        .into_iter()
-        .filter_map(|k| map.remove_entry(&k))
-        .collect()
 }
